Check rest of tools

+++ src/js/actions/store/image-store.js
@@ -1,8 +1,22 @@
-import { get } from "jquery";
+import { v4 as uuidv4 } from 'uuid';
 
-let idCounter = 0;
+// Get a unique id to identify this tab's history in the database
+let tabUuid;
+try {
+    tabUuid = sessionStorage.getItem('history_tab_uuid');
+} catch (error) {}
+if (!tabUuid) {
+    tabUuid = uuidv4();
+    try {
+        sessionStorage.setItem('history_tab_uuid', tabUuid);
+    } catch (error) {}
+}
+
+let imageIdCounter = 0;
 let database = null;
 let databaseInitPromise = null;
+const tabPingInterval = 60000;
+const assumeTabIsClosedTimeout = 300000; // Inactive tabs setInterval is slowed down in most browsers, this should be significantly higher than tabPingInterval
 
 export default {
     /**
@@ -13,16 +27,23 @@ export default {
             databaseInitPromise = new Promise(async (resolveInit) => {
                 try {
                     if (window.indexedDB) {
-                        // Delete database from a previous page load
-                        await new Promise((resolve, reject) => {
-                            let deleteRequest = window.indexedDB.deleteDatabase('undoHistoryImageStore');
-                            deleteRequest.onerror = () => {
-                                reject(deleteRequest.error);
-                            };
-                            deleteRequest.onsuccess = () => {
-                                resolve();
-                            };
-                        });
+                        // Delete database from a previous page load, if no other tabs have notified that they're open in a while
+                        let shouldDeleteDatabase = true;
+                        try {
+                            let lastDatabaseTabPing = localStorage.getItem('history_usage_ping');
+                            shouldDeleteDatabase = (!lastDatabaseTabPing || parseInt(lastDatabaseTabPing, 10) < new Date().getTime() - assumeTabIsClosedTimeout);
+                        } catch (error) {}
+                        if (shouldDeleteDatabase) {
+                            await new Promise((resolve, reject) => {
+                                let deleteRequest = window.indexedDB.deleteDatabase('undoHistoryImageStore');
+                                deleteRequest.onerror = () => {
+                                    reject(deleteRequest.error);
+                                };
+                                deleteRequest.onsuccess = () => {
+                                    resolve();
+                                };
+                            });
+                        }
                         // Initialize database
                         await new Promise((resolve, reject) => {
                             let openRequest = window.indexedDB.open('undoHistoryImageStore', 1);
@@ -45,6 +66,15 @@ export default {
                         if (!database) {
                             throw new Error('indexedDB not initialized');
                         }
+                        // Delete history from previous session
+                        try {
+                            await this.delete_all();
+                        } catch (error) {}
+                        // Ping localStorage for as long as this browser tab is open
+                        localStorage.setItem('history_usage_ping', new Date().getTime() + '');
+                        setInterval(() => {
+                            localStorage.setItem('history_usage_ping', new Date().getTime() + '');
+                        }, tabPingInterval);
                     }
                 } catch (error) {
                     database = {
@@ -68,7 +98,7 @@ export default {
      */
     async add(imageData) {
         await this.init();
-        let imageId = (idCounter++) + '';
+        let imageId = tabUuid + '-' + (imageIdCounter++);
         if (database.isMemory) {
             database.images[imageId] = imageData;
         } else {
@@ -77,6 +107,7 @@ export default {
                 const images = transaction.objectStore('images');
                 const image = {
                     id: imageId,
+                    tabUuid,
                     data: imageData
                 }
                 const request = images.add(image);
@@ -107,7 +138,7 @@ export default {
                 const images = transaction.objectStore('images');
                 const request = images.get(imageId);
                 request.onsuccess = function() {
-                    resolve(request.result.data);
+                    resolve(request.result && request.result.data);
                 };
                 request.onerror = function() {
                     reject(request.error);
@@ -120,7 +151,7 @@ export default {
      * Deletes the specified image from the database, by imageId retrieved from "add()" method.
      * 
      * @param {string} imageId the id of the image to delete
-     * @returns {Promise<string | canvas | ImageData>} resolves with the image
+     * @returns {Promise<void>} 
      */
     async delete(imageId) {
         await this.init();
@@ -135,6 +166,54 @@ export default {
                     resolve();
                 };
                 request.onerror = function() {
+                    reject(request.error);
+                };
+            });
+        }
+    },
+
+    /**
+     * Deletes all images associated with the current tab.
+     * 
+     * @returns {Promise<void>} 
+     */
+    async delete_all() {
+        await this.init();
+        if (database.isMemory) {
+            database.images = {};
+        } else {
+            return new Promise((resolve, reject) => {
+                const transaction = database.transaction('images', 'readwrite');
+                const images = transaction.objectStore('images');
+                const getAllImagesRequest = images.getAll();
+                getAllImagesRequest.onsuccess = async function () {
+                    const allImages = getAllImagesRequest.result;
+                    let errorOccurred = false;
+                    for (let image of allImages) {
+                        if (image.tabUuid === tabUuid) {
+                            try {
+                                await new Promise((deleteResolve, deleteReject) => {
+                                    const request = images.delete(image.id);
+                                    request.onsuccess = function() {
+                                        deleteResolve();
+                                    };
+                                    request.onerror = function() {
+                                        deleteReject(request.error);
+                                    };
+                                });
+                            } catch (error) {
+                                errorOccurred = true;
+                                // Should eventually be deleted when database is deleted due to timeout
+                            }
+                        }
+                    }
+                    if (errorOccurred) {
+                        // Use a different uuid to prevent conflicts
+                        tabUuid = uuidv4();
+                    }
+                    resolve();
+                };
+                getAllImagesRequest.onerror = function () {
                     reject(request.error);
                 };
             });
